+++ app/src/main/java/com/techcourse/HandlerMappings.java
@@ -1,7 +1,7 @@
 package com.techcourse;
 
 import jakarta.servlet.http.HttpServletRequest;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.HandlerMapping;
 
 import java.util.ArrayList;
 import java.util.List;
